requirements.txt
streamlit==1.52.1
pandas==2.3.3
plotly==6.3.1
openpyxl==3.1.5
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
python-dotenv==1.1.1
